requirements.txt
python-dotenv
pydantic
openai
streamlit
pandas
